requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
anthropic==0.42.0
PyJWT==2.10.1
pydantic==2.10.4
httpx==0.28.1
cryptography==44.0.0
